empty-demo/empty-agent/pyproject.toml
[project]
name = "empty-agent"
version = "0.1.0"
description = "Teams agent with OBO flow to Graph API and custom Empty API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "microsoft-agents-hosting-aiohttp>=0.6.1",
    "microsoft-agents-authentication-msal>=0.6.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]
